No longer disallow recursive call to get_config (reverts e390cbf)
It turns out that we make use of recursive `get_config` calls in our
own code: `CachedUpdatePersister` initializes with
`self.update_cache()`, which is decorated with
`PluggableDecorator('update_model_decorators')`, which calls
`get_config` to get a list of strings pointing to decorator functions.

In any event, it may be a good idea to call `self.update_cache()` from
a thread so to not block initialization.  However,
`PluggableDecorator` likely has other legitimate uses, so this isn't a
fix for the problem at hand.

The caveat still applies that if one calls `get_config` from component
code, one must be aware that the part of config we want to refer to
may not be initialized yet.  Best way to work around this is to either
not rely on `__factory__` and friends at all (PluggableDecorator does
that), or to make that part of the configuration a sub-entry to the
component's own configuration, if possible.