Decorate all ModelPersister read() and write() methods automatically
This removes the need to add the decorator in every class
individually, which is something that's easy to forget.

A side effect of this change is that for persisters that wrap other
persisters (such as the CachedUpdatePersister), the write decorator
will be called twice, since now both CachedUpdatePersister and the
underlying wrapped persister will fire and call the decorators.  I'm
considering this negligible; the effect is only felt with slow and non
idempotent code.